### 2025-01-17 #7623 @crvernon
:wave: @jbytecode - are you able to take this one on as editor?
