### 2025-01-17 #7623 @jbytecode
@crvernon - Sure, gladly! (I'm a little bit overloaded but I think I can multiple-dispatch myself for this submission!)
