### 2025-01-17 #7623 @jbytecode
@Oddan - Installation instructions are also missing in the repository. Could you please add a section in both readme.md and the documentation about how to install the package? It seems the package has already been registered in the Julia package registry, so it can be installed using

```julia
julia> using Pkg
(@1.11) pkg> add SurfaceWaterIntegratedModeling
```

but it can also be installed using the GitHub cloning. Please add them in proper places in docs.

Thank you in advance.
