### 2025-01-19 ac5c489
Added missing DOIs, and updated [compat] section of examples
