+++ docs/JOSS/paper.bib
@@ -15,7 +15,7 @@ volume = {8},
 pages = {100060},
 year = {2020},
 issn = {2589-9155},
-doi = {https://doi.org/10.1016/j.hydroa.2020.100060},
+doi = {10.1016/j.hydroa.2020.100060},
 url = {https://www.sciencedirect.com/science/article/pii/S2589915520300110},
 author = {Thomas Skaugen and Deborah Lawrence and Rengifo Zenon Ortega},
 }
@@ -26,6 +26,7 @@ author = {Thomas Skaugen and Deborah Lawrence and Rengifo Zenon Ortega},
   journal={IOP Conference Series: Earth and Environmental Science},
   year={2019},
   volume={304},
+  doi={10.1088/1755-1315/304/2/022088},
   url={https://api.semanticscholar.org/CorpusID:204259188}
 }
 
@@ -51,7 +52,7 @@ volume = {28},
 number = {15},
 pages = {4529-4542},
 keywords = {rainfall-runoff model, overparameterization, recession},
-doi = {https://doi.org/10.1002/hyp.9968},
+doi = {10.1002/hyp.9968},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/hyp.9968},
 year = {2014}
 }
@@ -98,7 +99,7 @@ volume = {118},
 number = {4},
 pages = {2105-2123},
 keywords = {flow routing, overland flow, modeled landscape, benchmark, process-transition, topographic metrics},
-doi = {https://doi.org/10.1002/jgrf.20127},
+doi = {10.1002/jgrf.20127},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/jgrf.20127},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/jgrf.20127},
 year = {2013}
