### 2025-01-19 7b08dc0
Added [compat] section of examples; removed ArchGDAL
Added [compat] section of the examples package folder.
Removed ArchGDAL from the SurfaceWaterIntegratedModeling package.
