+++ examples/Project.toml
@@ -1,5 +1,4 @@
 [deps]
-AbstractPlotting = "537997a7-5e4e-5d89-9595-2241ea00577e"
 ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
@@ -8,3 +7,12 @@ Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 SurfaceWaterIntegratedModeling = "71e2a337-b4d4-4738-9f2b-3c37e1d3d462"
+
+[compat]
+ArchGDAL = "0.10.2"
+ColorSchemes = "3.24"
+GLMakie = "0.10.18"
+Makie = "0.21"
+GeometryBasics = "0.4.11"
+Graphs = "1.9"
+Images = "0.26.1"
