### 2025-01-19 fcef755
Updated installation instructions to documetation
