### 2025-01-19 #7623 @Oddan
@jbytecode - Hi, thank you for your message.  I have now fixed the DOI issues, updated the text in readme.md and the documentation regarding installation, as requested.  I do not have any particular suggestion for reviewers, although the authors or editors involved in the JOSS paper [WDPM: the Wetland DEM Ponding Model](https://github.com/openjournals/joss-reviews/issues/2276) (identified by editorialbot above) might be a good choice, as it looks to me that a similar methodology has been used.
