### 2025-01-19 #7623 @jbytecode
@Oddan - It would be nice to have this part in the readme.md, too.

https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl/commit/fcef75533208a337cecafcb70ae387798e3df44a#diff-36c852523b6a7513a10b8a6bae21f1037eaa27484efc9042f75bbda46d9f529bR57
