### 2025-01-19 4528208
Additional info on installation in README.md
Also fixed a small orthographic error in documentation.
