### 2025-01-19 #7623 @Oddan
I added the package installation instruction from the documentation to the README.md file as well.
