### 2025-01-26 #7623 @pritamd47
> 👋👋👋 Dear @pritamd47 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software.
>
> The submission I'd like you to review is titled:
>
> **[[PRE REVIEW]: Topography-based surface water modeling in Julia, with support for infiltration and temporal developments](https://github.com/openjournals/joss-reviews/issues/7623)**
>
> You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7623).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
> Thank you in advance!

Hi @jbytecode. Thanks for the invite to review the submission. I am swamped at the moment with deadlines so I won't be able to review the paper unfortunately.
