### 2025-01-30 #7623 @jbytecode
👋👋👋 Hi @puruckertom, @tukss 👋👋👋

JOSS publishes articles about open source research software.

The submission I'd like you to review is titled:

**[[PRE REVIEW]: Topography-based surface water modeling in Julia, with support for infiltration and temporal developments](https://github.com/openjournals/joss-reviews/issues/7623)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7623).

Thank you in advance!
