### 2025-02-17 #7623 @jabirali
@jbytecode: As long as I can start my review process at the end of March, then I’d be happy to review the paper.

**Full disclosure:** Looking closer at the submission, I see that I actually know one of the authors from before (we have collaborated on a project before, but have not published any papers together). Is that considered a conflict of interest?
