### 2025-02-17 #7623 @jabirali
I think this might perhaps count as a COI; we collaborated on a grant application that is currently under consideration, which means that «recent (past four years) collaborators in funded research» might soon apply. Perhaps it’s best to find another referee to ensure neutrality :)
