### 2025-02-18 #7785 @editorialbot
[REVIEW]: Topography-based surface water modeling in Julia, with support for infiltration and temporal developments
**Submitting author:** @Oddan (<a href="http://orcid.org/0000-0002-2245-9512">Odd Andersen</a>)
**Repository:** https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl.git
**Branch with paper.md** (empty if default branch): JOSS_article
**Version:** v1.0.3
