### 2025-03-17 #7785 @tukss
@Oddan In the summary section the sentence "Spill-point analyses are highly computationally efficient compared to tools based on numerical simulation." is repeated.
