### 2025-03-17 #7785 @tukss
Here is my review:

The paper is in very good shape already and I only have some minor points to be corrected listed below. Please note that I am not a domain expert in hydrology, so while I verified that the promised functionality is provided and looks plausible, I cannot speak to its accuracy in detail.

# Documentation

-   typo: surfaecs <https://sintefmath.github.io/SurfaceWaterIntegratedModeling.jl/dev/>
-   The code has automated tests run by the CI but it would be good if the steps to run these were briefly described in the documentation
-   missing guidelines for community contribution

# Paper

-   repeated sentence in summary
-   typo in "Infiltration and temporal evolution": "and and extact" -> "and exact"
-   syntax of citing multiple references should be changed: e.g., `[@Beven:2003], [@Zhang:2019]` should be `[@Beven:2003; @Zhang:2019]`
