### 2025-03-20 53b02c9
Minor fixes pointed out by reviewer
