### 2025-03-21 06ab51d
Added instruction in the docs on how to run tests manually
