+++ CONTRIBUTING.md
@@ -1,0 +1,37 @@
+Contributing to SurfaceWaterIntegratedModeling.jl
+
+Thank you for considering contributing to this project! Here are some guidelines to help you get started.
+
+## Getting Started
+
+1. Fork the repository on GitHub.
+2. Clone your fork to your local machine:
+   ```sh
+   git clone git@github.com:your-username/SurfaceWaterIntegratedModeling.jl.git
+3. Navigate to the project directory:
+   cd SurfaceWaterIntegratedModeling.jl
+4. Install the required dependencies:
+   using Pkg
+   Pkg.instantiate()
+
+## Making Changes
+
+1. Create a new branch for your feature or bugfix:
+   git checkout -b my-feature-branch
+2. Make your changes and commit them with a clear and descriptive message:
+   git commit -am "Add new feature"
+3. Push your branch to your fork:
+   git push origin my-feature-branch
+
+## Running Tests
+
+1. Ensure all tests pass before submitting your changes:
+   using Pkg
+   Pkg.test()
+
+## Submitting a Pull Request
+
+1. Open a pull request on GitHub from your feature branch to the main branch of the original repository.
+2. Provide a clear description of your changes and any relevant information.
+
+Thank you for your contributions!
