### 2025-03-21 d7d0b98
Added 'sintefmath' to CONTRIBUTING.md
