### 2025-03-22 #7785 @Oddan
> # Documentation
>
>     * typo: surfaecs https://sintefmath.github.io/SurfaceWaterIntegratedModeling.jl/dev/
>
>     * The code has automated tests run by the CI but it would be good if the steps to run these were briefly described in the documentation
>
>     * missing guidelines for community contribution
>
>
> # Paper
>
>     * repeated sentence in summary
>
>     * typo in "Infiltration and temporal evolution": "and and extact" -> "and exact"
>
>     * syntax of citing multiple references should be changed: e.g., `[@Beven:2003], [@Zhang:2019]` should be `[@Beven:2003; @Zhang:2019]`

@tukss Thank you for pointing out the issues above. I've now checked in some changes to address them.
