### 2025-04-25 b364147
Slightly extended CONTRIBUTING guidelines
