### 2025-04-25 #7785 @Oddan
Thank you for the feedback!  I have now extended the CONTRIBUTION.md file with a short paragraph on using the issue tracker.
