### 2025-04-27 #7785 @mkitti
All the basics look good to me. I think this is ready for publication.

# Minor Points

* I usually like to see a small code snippet in the paper itself as part of the examples, but there are plenty of examples in the documentation.
* Contribution guidelines are not linked to from either the README or the documentation
