### 2025-04-27 #7785 @mkitti
I did a test of installing this package into an fresh Julia depot and noticed that this package has more than 400 indirect dependencies and requires the precompilation of 517 packages and extensions to install. I'm wondering if this could be made easier by converting some dependencies to optional dependencies.

In examining the package structure, I believe GLMakie could be dropped as a dependency. There is nothing explicitly importing GLMakie in the package.
https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl/pull/2

I wonder if DifferentialEquations.jl could be made into an optional dependency.
