### 2025-04-27 ef9c436
Drop Images.jl and GraphPlot.jl as dependencies, update compat
