+++ Project.toml
@@ -8,11 +8,8 @@ ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
-GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 LazyArtifacts = "4af54fe1-eca0-43a8-85a7-787d91b784e3"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
@@ -22,15 +19,12 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 ColorSchemes = "3.24"
-Colors = "0.12.10"
+Colors = "0.12.10, 0.13"
 DifferentialEquations = "7.13"
-GLMakie = "0.10.18"
-GeometryBasics = "0.4.11"
-GraphPlot = "0.6"
+GeometryBasics = "0.4.11, 0.5"
 Graphs = "1.9"
-Images = "0.26.1"
 Interpolations = "0.15.1"
-Makie = "0.21"
+Makie = "0.21, 0.22"
 OffsetArrays = "1.13"
 Roots = "2.1.2"
 julia = "1.9"
