### 2025-04-27 #7785 @mkitti
It looks like we can drop Images.jl and GraphPlots.jl:
https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl/pull/4

We can also remove all deprecations if we update compat to GeometryBasics.jl v0.5:
https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl/pull/5
