### 2025-04-27 #7785 @Oddan
@mkitti   Thanks for your suggestions.  I have added a link to the contribution guidelines from the README.md file.  I also agree that the dependency situation could be improved.  In that regard I highly appreciate your pull requests, and will look into them as soon as I can.
