### 2025-04-28 9166236
add citation to Julia
