### 2025-04-28 #7785 @jbytecode
@Oddan - I've just sent a pull request for both the manuscript and the bibtex file:

https://github.com/sintefmath/SurfaceWaterIntegratedModeling.jl/pull/6

Please review the changes. If you are agree with me, please merge.

Please have a full read of the paper and correct anything including typos, if exists.

Please ping me when you've done with them.

Thank you in advance.
