+++ Project.toml
@@ -21,7 +21,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 ColorSchemes = "3.24"
 Colors = "0.12.10, 0.13"
 DifferentialEquations = "7.13"
-GeometryBasics = "0.5"
+GeometryBasics = "0.4.11"
 Graphs = "1.9"
 Interpolations = "0.15.1"
 Makie = "0.21, 0.22"
