### 2025-04-29 c97c230
Slight changes to dependencies and compat
