### 2025-04-30 #7785 @jbytecode
@Oddan - The archive metadata in Zenodo should match the ones in the submission, e.g. the title of the manuscript, the author names, affiliations, and ORCIDs. It seems the title doesn't match. Please fix the title and ping me again.
