### 2025-04-30 #7785 @Oddan
But it is the project source code that is stored at Zenodo, not the journal article. As such, shouldn't it have the project name? It doesn't seem to make sense to me to give the software the name of the journal article.

Or did I misunderstand what you meant?

Sent with [Proton Mail](https://proton.me/mail/home) secure email.
