### 2025-04-30 #7785 @jbytecode
This is not the standards of Zenodo, this is about the journal:

https://joss.readthedocs.io/en/latest/sample_messages.html#message-to-authors-at-the-end-of-a-review

`- Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.`

Please, follow the instructions that I listed.
