### 2025-04-30 #7785 @jbytecode
@Oddan - Looks fine to me. I'm now recommending an acceptance. Our track editor will make the final decision. Thank you again for submitting your work to JOSS.
