### 2025-04-30 #7785 @crvernon
:wave: @Oddan  - The submission is looking great.

One question for you to address:

- [ ] In the figure 4 caption, did you mean "river" instead of "rive" ?

Additionally, I need you to fix the following:

- [ ] LINE 113:  "U.s. Geological survey" should be written "United States Geological Survey (USGS)"
- [ ] LINE 126:  "jialingjiang river basin" should be written "Jialingjiang River Basin"

You can maintain capitalization in your bib file using curly brackets around the characters you wish to maintain the formatting of.

Let me know when these have been taken care of and I'll accept.  Thank you!
