+++ docs/JOSS/paper.bib
@@ -21,7 +21,7 @@ author = {Thomas Skaugen and Deborah Lawrence and Rengifo Zenon Ortega},
 }
 
 @article{Zhang:2019,
-  title={Preliminary Research and Application of MIKE SHE Model in Jialingjiang River Basin},
+  title={Preliminary Research and Application of MIKE SHE Model in {J}ialingjiang {R}iver {B}asin},
   author={Meng-Di Zhang and Jing Zhang and Yongyu Song},
   journal={IOP Conference Series: Earth and Environmental Science},
   year={2019},
@@ -32,7 +32,7 @@ author = {Thomas Skaugen and Deborah Lawrence and Rengifo Zenon Ortega},
 
 @misc{Langevin:2017,
   author= {Langevin, C.D., Hughes, J.D., Banta, E.R., Provost, A.M., Niswonger, R.G., Panday, Sorab},
-  title = {MODFLOW 6 Modular Hydrologic Model: U.S. Geological Survey Software},
+  title = {MODFLOW 6 Modular Hydrologic Model: {U}nited {S}tates {G}eological {S}urvey (USGS) Software},
   year= {2017},
   url= {https://doi.org/10.5066/F76Q1VQV},
 }
